add relacoes-tabelas

+++ backend/src/controllers/produtos.js
@@ -4,15 +4,15 @@ const Produto = models.Produto;
 const Op = Sequelize.Op;
 import {v4 as uuidv4} from 'uuid';
 
-exports.index = async(req,res) => {
+const index = async(req,res) => {
     const produtos = await Produto.findAll({});
     res.status(200).send(produtos);
 };
 
-exports.create = async(req,res) => {
+const create = async(req,res) => {
     try{
         let produto = await Produto.create({
-            id: uuidv4(),
+            //id: uuidv4(),
             ...req.body
         });
         res.status(201).send({
@@ -25,40 +25,54 @@ exports.create = async(req,res) => {
     }
 };
 
-exports.read = async(req,res) => {
+const read = async(req,res) => {
     const {id} = req.params;
     try{
         const produto = await Produto.findByPk(id);
-        res.status(200).send({produto});
+        if(produto){
+            res.status(200).send({produto});
+        }else{
+            res.status(204).send();
+        }
     }catch(e){
         console.log(e.errors);
         res.status(500).send({error: e});
     }
 };
 
-exports.update = async(req,res) => {
+const update = async(req,res) => {
     const produto = await Produto.findOne({where : {id: req.params.id}});
-    try{
-        const produto_update = await Produto.update({
-            ...req.body
-        }, { where : {id: req.params.id}});
-        
-        res.status(200).send({produto_update});
-    }catch(e){
-        console.log(e.errors);
-        res.status(500).send({error: e});
+    if(produto){
+        try{
+            const produto_update = await Produto.update({
+                ...req.body
+            }, { where : {id: req.params.id}});
+            
+            res.status(200).send({produto_update});
+        }catch(e){
+            console.log(e.errors);
+            res.status(500).send({error: e});
+        }
+    }else{
+        res.status(204).send();
     }
 };
 
-exports.remove = async(req,res) => {
+const remove = async(req,res) => {
    const {id} = req.params;
-   try{
-        await Produto.destroy({where: {id: id}});
-        res.status(200).send({message:'Produto removido com sucesso!'});  
-   }catch(e){
-       console.log(e);
-       res.status(500).send({error: e});
-   }
+   const produto = await Produto.findOne({where : {id: req.params.id}});
+   if(produto){
+        try{
+                await Produto.destroy({where: {id: id}});
+                res.status(200).send({message:'Produto removido com sucesso!'});  
+        }catch(e){
+            console.log(e);
+            res.status(500).send({error: e});
+        }
+    }else{
+        res.status(204).send();
+    }
 };
 
+export default { index, create, read, update, remove };
